Move abstract primitive classes into primitive package.
